Add a users endpoint and an install subcommand
Add a users endpoint that can be used by admins to GET information on
all users and can also be used to PUT information on a user.
Add an install subcommand that can be used to install the Charity Showcase
Server project onto the user's machine. This currently only creates the
database and populates it with test data. In future this subcommand may
have options for installing a test, development, or production
environment.
Add testdata.json that can be used to populate the database with example
data.